requirements.txt
dj-database-url==0.4.1
Django==1.11.5
djangorestframework==3.7.0
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.3.1
beautifulsoup4==4.6.0
requests==2.18.4
html5lib==0.999999999
